Call mouse event directly without event queue
