+++ luda-editor/client/src/pages/sequence_edit_page/loaded/components/wysiwyg_editor/update/mod.rs
@@ -49,18 +49,19 @@ impl WysiwygEditor {
                 InternalEvent::MouseDownContainer => {
                     self.editing_image_index = None;
                 }
-            });
-
-        if let Some(NamuiEvent::MouseUp(event)) = event.downcast_ref() {
-            if let Some(Dragging::Mover { context }) = self.dragging.as_mut() {
-                context.end_global_xy = event.xy;
-                if let Some(index) = self.editing_image_index {
-                    let circumscribed = self.screen_images[index].as_ref().unwrap().circumscribed;
-                    let moved_circumscribed = context.move_circumscribed(circumscribed);
-                    self.screen_images[index].as_mut().unwrap().circumscribed = moved_circumscribed;
+                &InternalEvent::MouseUp { global_xy } => {
+                    if let Some(Dragging::Mover { context }) = self.dragging.as_mut() {
+                        context.end_global_xy = global_xy;
+                        if let Some(index) = self.editing_image_index {
+                            let circumscribed =
+                                self.screen_images[index].as_ref().unwrap().circumscribed;
+                            let moved_circumscribed = context.move_circumscribed(circumscribed);
+                            self.screen_images[index].as_mut().unwrap().circumscribed =
+                                moved_circumscribed;
+                        }
+                        self.dragging = None;
+                    }
                 }
-                self.dragging = None;
-            }
-        }
+            });
     }
 }
