+++ namui-animation-editor/src/time_point_editor/timeline_window/render/mod.rs
@@ -13,8 +13,8 @@ impl TimelineWindow {
                         .on_wheel(move |event| {
                             let mouse_global_xy = namui::mouse::position();
                             let table_xy = event
-                                .namui_context
-                                .get_rendering_tree_xy(event.target)
+                                .root
+                                .get_xy_of_child(event.target)
                                 .expect("ERROR: fail to get rendering_tree_xy");
 
                             let mouse_local_xy = mouse_global_xy - table_xy;
