+++ namui-animation-editor/src/time_point_editor/wysiwyg_window/render/layer/editing_tool/rotate.rs
@@ -81,10 +81,7 @@ impl WysiwygWindow {
             let window_id = self.window_id.clone();
             let real_px_per_screen_px = self.real_px_per_screen_px;
             builder.on_mouse_down_in(move |event| {
-                let window_global_xy = event
-                    .namui_context
-                    .get_rendering_tree_xy_by_id(window_id)
-                    .unwrap();
+                let window_global_xy = event.root.get_xy_by_id(window_id).unwrap();
                 let mouse_local_xy = (event.global_xy - window_global_xy) * real_px_per_screen_px;
 
                 namui::event::send(Event::RotationToolMouseDown {
