+++ namui-animation-editor/src/time_point_editor/wysiwyg_window/render/mod.rs
@@ -34,8 +34,8 @@ impl WysiwygWindow {
                     .on_wheel(move |event| {
                         let mouse_global_xy = namui::mouse::position();
                         let row_xy = event
-                            .namui_context
-                            .get_rendering_tree_xy(event.target)
+                            .root
+                            .get_xy_of_child(event.target)
                             .expect("ERROR: fail to get rendering_tree_xy");
 
                         let mouse_local_xy = mouse_global_xy - row_xy;
