+++ namui/src/namui/namui_context/mod.rs
@@ -42,10 +42,4 @@ impl NamuiContext {
         crate::graphics::request_animation_frame(on_frame);
         self.run_main_loop(state, props).await;
     }
-    pub fn get_rendering_tree_xy_by_id(&self, id: crate::Uuid) -> Option<Xy<Px>> {
-        self.rendering_tree.get_xy_by_id(id)
-    }
-    pub fn get_rendering_tree_xy(&self, rendering_tree: &RenderingTree) -> Option<Xy<Px>> {
-        self.rendering_tree.get_xy_of_child(rendering_tree)
-    }
 }
